refactor(cpg_flow_test/): move all test code into one folder
