--- test.py
@@ -1,10 +1,0 @@
-json = {
-    'COH2142': {
-        'CPG280370': 'gs://cpg-fewgenomes-test/CPG280370_no_evens.txt',
-        'CPG280388': 'gs://cpg-fewgenomes-test/CPG280388_no_evens.txt',
-        'CPG280396': 'gs://cpg-fewgenomes-test/CPG280396_no_evens.txt',
-        'CPG280404': 'gs://cpg-fewgenomes-test/CPG280404_no_evens.txt',
-        'CPG280412': 'gs://cpg-fewgenomes-test/CPG280412_no_evens.txt',
-        'no_evens': 'gs://cpg-fewgenomes-test/prime_pyramid/COH2142_no_evens.txt',
-    },
-}
